Title: 두 수의 합, Time: 86ms, Memory: 5MB, Status: Wrong Answer - Codetree

+++ 250720/두 수의 합/sum-of-two-num.cpp
@@ -16,7 +16,7 @@ int main() {
     }
 
     for(auto it : m){
-        int tmp = k - it.first;
+        long long tmp = k - it.first;
 
         if(tmp == it.first){
             ret += (it.second *(it.second-1))/2;
